@@ -1,6 +1,6 @@
 [metadata]
 name = identify
-version = 1.5.13
+version = 2.1.1
 description = File identification library for Python
 long_description = file: README.md
 long_description_content_type = text/markdown
@@ -11,26 +11,26 @@ license = MIT
 license_file = LICENSE
 classifiers =
     License :: OSI Approved :: MIT License
-    Programming Language :: Python :: 2
-    Programming Language :: Python :: 2.7
     Programming Language :: Python :: 3
-    Programming Language :: Python :: 3.4
-    Programming Language :: Python :: 3.5
+    Programming Language :: Python :: 3 :: Only
     Programming Language :: Python :: 3.6
     Programming Language :: Python :: 3.7
+    Programming Language :: Python :: 3.8
+    Programming Language :: Python :: 3.9
     Programming Language :: Python :: Implementation :: CPython
     Programming Language :: Python :: Implementation :: PyPy
 
 [options]
 packages = find:
-python_requires = >=2.7, !=3.0.*, !=3.1.*, !=3.2.*, !=3.3.*
+python_requires = >=3.6.1
 
 [options.entry_points]
 console_scripts =
     identify-cli=identify.cli:main
 
 [options.extras_require]
-license = editdistance
+license =
+    editdistance
 
 [options.packages.find]
 exclude =
@@ -42,3 +42,16 @@ universal = True
 
 [coverage:run]
 plugins = covdefaults
+
+[mypy]
+check_untyped_defs = true
+disallow_any_generics = true
+disallow_incomplete_defs = true
+disallow_untyped_defs = true
+no_implicit_optional = true
+
+[mypy-testing.*]
+disallow_untyped_defs = false
+
+[mypy-tests.*]
+disallow_untyped_defs = false
